myproject/requirements.txt
asgiref==3.8.1
bson==0.5.8
dataclasses==0.6
Django==2.2.28
django-cors-headers==3.7.0
djangorestframework==3.12.4
djongo==1.3.6
paho-mqtt==1.5.1
pymongo==3.12.3
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
sqlparse==0.2.4
typing_extensions==4.11.0
tzdata==2024.1
